ip/requirements.txt
vunit_hdl

# packages for TerosHDL
teroshdl
cocotb
yowasp-yosys
edalize
vsg
